Implement `GetOrElse` on `Either` with Tests
