+++ Funcky.Test/Monads/EitherTest.cs
@@ -1,6 +1,5 @@
 using FsCheck;
 using FsCheck.Xunit;
-using Xunit.Sdk;
 
 namespace Funcky.Test.Monads;
 
@@ -114,31 +113,6 @@ public sealed partial class EitherTest
 
         Assert.False(hasLeft);
         Assert.True(hasRight);
-    }
-
-    [Fact]
-    public void InspectDoesNothingWhenEitherIsLeft()
-    {
-        var either = Either<string, int>.Left("foo");
-        either.Inspect(_ => throw new XunitException("Side effect was unexpectedly called"));
-    }
-
-    [Fact]
-    public void InspectCallsSideEffectWhenEitherIsRight()
-    {
-        const int value = 10;
-        var either = Either<string>.Return(value);
-
-        var sideEffect = Option<int>.None;
-        either.Inspect(v => sideEffect = v);
-        FunctionalAssert.IsSome(value, sideEffect);
-    }
-
-    [Theory]
-    [MemberData(nameof(LeftAndRight))]
-    public void InspectReturnsOriginalValue(Either<string, int> either)
-    {
-        Assert.Equal(either, either.Inspect(NoOperation));
     }
 
     public static TheoryData<Either<string, int>> LeftAndRight()
